fix windows clients disconnecting en masse
When a client closed it would signal an event to tell the worker thread
to stop processing enet messages, however, all windows clients on the
same machine were listening to the exact same event, causing all windows
clients to hang and be forcefully disconnected by the server.

Fix this by creating events without a name so they are all unique to
each client.

Also added a "-d" debug param.